lambda_poller/requirements.txt
snowflake-connector-python==3.7.0
boto3==1.34.44